+++ packages/react/src/atomic/Link/src/link.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react/require-default-props */
 import React, { forwardRef } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { ExternalLinkIcon } from "@nimbus-ds/icons";
@@ -6,12 +7,22 @@ import { Text } from "@nimbus-ds/text";
 
 import { Link as LinkComponent, LinkProps } from "./Link";
 
-export const Basic: React.FC<LinkProps> = forwardRef(
-  ({ children = "Link", ...props }: LinkProps) => (
-    <LinkComponent {...props}>{children}</LinkComponent>
+export const Basic: React.FC<LinkProps & { disabled?: boolean }> = forwardRef(
+  ({
+    children = "Link",
+    disabled = false,
+    ...props
+  }: LinkProps & { disabled?: boolean }) => (
+    <LinkComponent {...props} as={disabled ? "button" : "a"}>
+      {children}
+    </LinkComponent>
   )
-) as React.FC<LinkProps>;
+) as React.FC<LinkProps & { disabled?: boolean }>;
+
 Basic.displayName = "Link";
+Basic.defaultProps = {
+  disabled: false,
+};
 
 const meta: Meta<typeof Basic> = {
   title: "Atomic/Link",
@@ -102,5 +113,13 @@ export const asButton: Story = {
   args: {
     children: "Link as button",
     as: "button",
+  },
+};
+
+export const disabled: Story & { args: LinkProps & { disabled: boolean } } = {
+  args: {
+    as: "button",
+    disabled: true,
+    children: "Link",
   },
 };