+++ packages/react/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Nimbus is an open-source Design System created by Tiendanube / Nuvemshop's team to empower and enhance more stories every day, with simplicity, accessibility, consistency and performance.
 
+## 2025-09-16 `5.23.0`
+
+#### 🎉 New features
+
+- `Box`: Added `ai-generative-interactive` and `ai-generative-interactive-border` options to `background` property, and `aiFocus` option to `boxShadow` property. ([#352](https://github.com/TiendaNube/nimbus-design-system/pull/352) by [@joacotornello](https://github.com/joacotornello))
+
+## 2025-09-15 `5.22.0`
+
+#### 🎉 New features
+
+- Added `ai-generative` appearance to `Textarea` component. ([#351](https://github.com/TiendaNube/nimbus-design-system/pull/351) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-09-15 `5.21.3`
 
 #### 🐛 Bug fixes